fix routes and controller message

+++ src/database/Mongo/Controllers/messageController.ts
@@ -1,5 +1,6 @@
 import { type Request, type Response } from 'express'
 import { JoiRequestValidatorInstance } from '../../../JoiRequestValidator'
+import Reactions from '../../../reactions'
 
 const Message = require('../Models/MessageModel')
 
@@ -39,13 +40,17 @@ async function getMessageById (req: Request, res: Response): Promise<Response> {
 async function deleteMessage (req: Request, res: Response): Promise<Response> {
   try {
     const { id } = req.params
-    const { deletedCount } = await Message.deleteOne({ _id: id })
-
-    if (deletedCount === 0) {
-      return res.status(400).send('Message to delete not found')
+    const message = await Message.findOne({ _id: id })
+    if (!message) {
+      return res.status(404).send('Message not found')
     }
 
-    return res.status(200).send('Message deleted successfully')
+    const { deletedCount } = await Message.deleteOne({ _id: id })
+    if (deletedCount === 0) {
+      return res.status(400).send('Message cannot be deleted')
+    }
+
+    return res.status(200).json({ message })
   } catch (error) {
     return res.status(500).json({ 'Internal Server Error': error })
   }
@@ -54,12 +59,16 @@ async function deleteMessage (req: Request, res: Response): Promise<Response> {
 async function editMessage (req: Request, res: Response): Promise<Response> {
   try {
     const { id } = req.params
-    const { editContent } = req.body
+    const { newMessageContent } = req.body
 
-    await Message.updateOne({ _id: id }, { content: editContent, edited: true })
+    const { error } = JoiRequestValidatorInstance.validate(req)
+    if (error) { return res.status(400).json({ error }) }
+
+    await Message.findOneAndUpdate({ _id: id }, { content: newMessageContent, edited: true })
     const message = await Message.findOne({ _id: id })
+    if (!message) { return res.status(404).send('Message not found') }
 
-    return res.status(200).send(message)
+    return res.status(200).send({ message })
   } catch (error) {
     return res.status(500).json({ 'Internal Server Error': error })
   }
@@ -71,12 +80,19 @@ async function reactToMessage (req: Request, res: Response): Promise<Response> {
     const { reaction } = req.body
     const { user } = req.body
 
-    const message = await Message.findById(id)
-    if (!message) {
+    const { error } = JoiRequestValidatorInstance.validate(req)
+    if (error) { return res.status(400).json({ error }) }
+
+    const messageToUpdate = await Message.findById(id)
+    if (!messageToUpdate) {
       return res.status(404).send('Message not found')
     }
 
-    const newReactionsMap = message.reactions || new Map()
+    if (!Reactions.includes(reaction)) {
+      return res.status(400).send('Reaction not found')
+    }
+
+    const newReactionsMap = messageToUpdate.reactions || new Map()
     newReactionsMap.set(user.id, reaction)
 
     const { modifiedCount } = await Message.updateOne({ _id: id }, { reactions: newReactionsMap })
@@ -85,7 +101,8 @@ async function reactToMessage (req: Request, res: Response): Promise<Response> {
       return res.status(400).send('Already reacted with this reaction')
     }
 
-    return res.status(200).json(message)
+    const message = await Message.findById(id)
+    return res.status(200).json({ message })
   } catch (error) {
     return res.status(500).json({ 'Internal Server Error': error })
   }
